Add concern to set downloadable headers
